requirements.txt
Package                       Version    
----------------------------- -----------
absl-py                       0.7.1      
affine                        2.2.2      
alabaster                     0.7.12     
astor                         0.8.0      
attrs                         19.1.0     
Babel                         2.7.0      
backcall                      0.1.0      
bleach                        3.1.0      
certifi                       2019.3.9   
chardet                       3.0.4      
Click                         7.0        
click-plugins                 1.1.1      
cligj                         0.5.0      
cycler                        0.10.0     
Cython                        0.29.10    
decorator                     4.4.0      
defusedxml                    0.6.0      
Django                        2.2.3      
docutils                      0.14       
entrypoints                   0.3        
fire                          0.2.1      
gast                          0.2.2      
grpcio                        1.21.1     
h5py                          2.9.0      
idna                          2.8        
image                         1.5.27     
imageio                       2.5.0      
imagesize                     1.1.0      
imgaug                        0.2.9      
imutils                       0.5.2      
ipdb                          0.12       
ipykernel                     5.1.1      
ipyparallel                   6.2.4      
ipython                       7.5.0      
ipython-genutils              0.2.0      
ipywidgets                    7.4.2      
jedi                          0.13.3     
Jinja2                        2.10.1     
joblib                        0.14.0     
json5                         0.9.1      
jsonschema                    3.0.1      
jupyter-client                5.2.4      
jupyter-core                  4.4.0      
jupyterlab                    1.2.6      
jupyterlab-server             1.0.6      
Keras                         2.2.4      
Keras-Applications            1.0.8      
Keras-Preprocessing           1.1.0      
kiwisolver                    1.1.0      
Markdown                      3.1.1      
MarkupSafe                    1.1.1      
mask-rcnn                     2.1        
matplotlib                    3.1.0      
mistune                       0.8.4      
mock                          3.0.5      
mrcnn                         0.2        
nbconvert                     5.5.0      
nbformat                      4.4.0      
networkx                      2.3        
nose                          1.3.7      
notebook                      5.7.8      
numpy                         1.16.4     
opencv-python                 4.1.0.25   
packaging                     19.0       
pandas                        0.24.2     
pandocfilters                 1.4.2      
parso                         0.4.0      
pexpect                       4.7.0      
pickleshare                   0.7.5      
Pillow                        6.0.0      
pip                           19.1.1     
prometheus-client             0.6.0      
prompt-toolkit                2.0.9      
protobuf                      3.8.0      
ptyprocess                    0.6.0      
pycocotools                   2.0        
Pygments                      2.4.2      
pyparsing                     2.4.0      
pyrsistent                    0.15.2     
python-dateutil               2.8.0      
pytz                          2019.1     
PyWavelets                    1.0.3      
PyYAML                        5.1        
pyzmq                         18.0.1     
qtconsole                     4.5.1      
rasterio                      1.0.24     
readchar                      2.0.1      
regex                         2020.1.8   
requests                      2.22.0     
savReaderWriter               3.4.2      
scikit-image                  0.15.0     
scikit-learn                  0.21.3     
scipy                         1.3.0      
seaborn                       0.9.0      
Send2Trash                    1.5.0      
setuptools                    41.0.1     
Shapely                       1.6.4.post2
six                           1.12.0     
sklearn                       0.0        
snowballstemmer               1.2.1      
snuggs                        1.4.6      
Sphinx                        2.1.0      
sphinxcontrib-applehelp       1.0.1      
sphinxcontrib-devhelp         1.0.1      
sphinxcontrib-htmlhelp        1.0.2      
sphinxcontrib-jsmath          1.0.1      
sphinxcontrib-qthelp          1.0.2      
sphinxcontrib-serializinghtml 1.1.3      
sqlparse                      0.3.0      
tensorboard                   1.13.1     
tensorflow                    1.13.1     
tensorflow-estimator          1.13.0     
termcolor                     1.1.0      
terminado                     0.8.2      
testpath                      0.4.2      
tornado                       6.0.3      
tqdm                          2.2.3      
traitlets                     4.3.2      
urllib3                       1.25.3     
wcwidth                       0.1.7      
webencodings                  0.5.1      
Werkzeug                      0.15.4     
wheel                         0.33.4     
widgetsnbextension            3.4.2      
